Move device page constants to common/constants
