+++ src/js/common/constants/dataVisualization.js
@@ -1,0 +1,9 @@
+export const VIEW_MODE = Object.freeze({
+  CARD: 'CARD',
+  TABLE: 'TABLE',
+});
+
+export const DATA_ORDER = Object.freeze({
+  ASC: 'asc',
+  DESC: 'desc',
+});
