+++ src/js/common/constants/index.js
@@ -1,1 +1,2 @@
 export { SOURCE, WIDGET, OPERATION } from './wizard';
+export { DATA_ORDER, VIEW_MODE } from './dataVisualization';
