--- src/js/views/devices/constants.js
@@ -1,4 +1,0 @@
-export const VIEW_MODE = {
-  CARD: 'CARD',
-  TABLE: 'TABLE',
-};
